feat: Dodata glavna struktura za handlovanje exception-a
- implementiran FigtherNotFoundException sa status kodom i porukom
- dodat ApiError DTO za osnovne error poruke

+++ src/main/java/org/stajkovic/ufc/cv/fighter/dto/FighterResponse.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
 import org.stajkovic.ufc.cv.fighter.model.Stance;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 // GET -> /ufc/fighters
@@ -16,22 +17,22 @@ public record FighterResponse(
         int wins,
         int losses,
         int draws,
-        double height,
-        double weight,
-        double reach,
+        BigDecimal height,
+        BigDecimal weight,
+        BigDecimal reach,
         Stance stance,
 
         @JsonFormat(pattern = "dd-MM-yyyy")
         LocalDate dateOfBirth,
-        
-        double significantStrikesLanded,
-        double significantStrikesAbsorbed,
-        double significantStrikesAccuracy,
-        double significantStrikesDeffence,
-        double avgTakedownsLanded,
-        double avgSubmissionsAttempted,
-        double takedownAccuracy,
-        double takedownDefense
+
+        BigDecimal significantStrikesLanded,
+        BigDecimal significantStrikesAbsorbed,
+        BigDecimal significantStrikesAccuracy,
+        BigDecimal significantStrikesDeffence,
+        BigDecimal avgTakedownsLanded,
+        BigDecimal avgSubmissionsAttempted,
+        BigDecimal takedownAccuracy,
+        BigDecimal takedownDefense
 ) {
 
 }
